Extracting get page details function.
